Marking tests with xfail, refactoring
Moved stuff around a little, and ensured that tests pass from now on.
