+++ qleet/__init__.py
@@ -1,3 +1,2 @@
 import qleet.circuits
-import qleet.plotter
-import qleet.pqc_property
+import qleet.analyzers
