--- qleet/plotter/__init__.py
@@ -1,1 +1,0 @@
-import qleet.plotter.landscape
